requirements.txt
travispy
github3.py